Creating a setup file
Sir this step is done on different systems like on mine it'll be different and on yours, it will be different. A PEC Device software installer will be installed on your computer or you can directly install or burn it on the device that you will experiment on as mass storage one.